// frontend/src/types/types.ts
enum SnippetEnum {
  allPlaylists,
  completedEvents,
  liveEvents,
  multipleChannels,
  multiplePlaylists,
  popularUploads,
  recentUploads,
  singlePlaylist,
  subscriptions,
  upcomingEvent,
}
// ...
export type ChannelType = {
  kind: string;
  etag: string;
  id: string;
  snippet: {
    title: string;
    description: string;
    customUrl: string;
    publishedAt: string;
    channelId: string;
    thumbnails: {
      default: {
        url: string;
        width: number;
        height: number;
      };
      medium: {
        url: string;
        width: number;
        height: number;
      };
      high: {
        url: string;
        width: number;
        height: number;
      };
    };
    defaultLanguage: string;
    localized: {
      title: string;
      description: string;
    };
    country: string;
  };
  contentDetails?: {
    relatedPlaylists: {
      likes: string;
      favorites: string;
      uploads: string;
    };
  };
  statistics?: {
    viewCount: string;
    subscriberCount: string;
    hiddenSubscriberCount: boolean;
    videoCount: string;
  };
  topicDetails?: {
    topicIds: string[];
    topicCategories: string[];
  };
  status?: {
    privacyStatus: string;
    isLinked: boolean;
    longUploadsStatus: string;
    madeForKids: boolean;
    selfDeclaredMadeForKids: boolean;
  };
  brandingSettings?: {
    channel?: {
      title: string;
      description: string;
      keywords: string;
      trackingAnalyticsAccountId: string;
      unsubscribedTrailer: string;
      defaultLanguage: string;
      country: string;
    };
    watch?: {
      textColor: string;
      backgroundColor: string;
      featuredPlaylistId: string;
    };
    image?: {
      bannerExternalUrl: string;
    };
  };
  auditDetails?: {
    overallGoodStanding: boolean;
    communityGuidelinesGoodStanding: boolean;
    copyrightStrikesGoodStanding: boolean;
    contentIdClaimsGoodStanding: boolean;
  };
  contentOwnerDetails?: {
    contentOwner: string;
    timeLinked: string;
  };
  localizations?: {
    key: {
      title: string;
      description: string;
    };
  };
};

export type ChannelInfoType = {
  kind: string;
  etag: string;
  nextPageToken?: string;
  prevPageToken?: string;
  pageInfo: {
    totalResults: number;
    resultsPerPage: number;
  };
  items: [ChannelType];
};
// ...
export type PlaylistItemType = {
  kind: string;
  etag: string;
  id: string;
  snippet: {
    publishedAt: string;
    channelId: string;
    title: string;
    description: string;
    thumbnails: {
      default: {
        url: string;
        width: number;
        height: number;
      };
      medium: {
        url: string;
        width: number;
        height: number;
      };
      high: {
        url: string;
        width: number;
        height: number;
      };
      standard: {
        url: string;
        width: number;
        height: number;
      };
      maxres: {
        url: string;
        width: number;
        height: number;
      };
    };
    channelTitle: string;
    videoOwnerChannelTitle: string;
    videoOwnerChannelId: string;
    playlistId: string;
    position: number;
    resourceId: {
      kind: string;
      videoId: string;
    };
  };
  contentDetails: {
    videoId: string;
    startAt: string;
    endAt: string;
    note: string;
    videoPublishedAt: number;
  };
  status: {
    privacyStatus: string;
  };
};

export type PlaylistItemListType = {
  kind: string;
  etag: string;
  nextPageToken: string;
  prevPageToken: string;
  pageInfo: {
    totalResults: number;
    resultsPerPage: number;
  };
  items: [PlaylistItemType];
};
// ...
export type ChannelSectionType = {
  kind: string;
  etag: string;
  items: [
    {
      kind: string;
      etag: string;
      id: string;
      snippet: {
        type: SnippetEnum;
        channelId: string;
        title: string;
        position: number;
      };
      contentDetails?: {
        playlists?: [string];
        channels?: [string];
      };
    }
  ];
};